[enhancement] - Scrolling performance issues
This commit will resolve a known scrolling performance issue related to scrollable areas.
The thing is that scrolling force repaints (draw the UI pixels) whenever the user is scrolling,
and that causes a lot of repaints that could be avoided using GPU accelerated repaints.

For legacy fix I'm using transform: translateZ(0) which will give the GPU to handle those repaints which make
the scrolling performance more better

Affected pages among the examples with this issue:
- http://react-component.github.io/select/examples/tags.html
- http://react-component.github.io/select/examples/single.html
- http://react-component.github.io/select/examples/multiple.html

Inspect the Chrome "rendering" tab in Devtools to validate the issue.
I'll be attaching a screenshots for the reproduction