fix: only update value with a good input value
A bad input value would be result in an undefined calculated value.
Validating an undefined value could empty the select control.
So when losing focus from a multiple select which already has some values,
the select would be emptied.

a bad input value should not be update the value to be validated.