Allow limiting number of items in the dropdown
In some use cases the number of options passed to rc-select can be very
large. The performance of rendering the dropdown with a large number of
items can be very poor. The dropdownLimit option if set will prevent
rendering of any matched options beyond the limit, greatly improving
performance. The user experience is not greatly diminished, as they are
presented with a smaller number of choices at once, and can narrow down
the choices by typing more, similar to a typeahead system.